+++ src/main/java/IdResponse.java
@@ -1,15 +1,15 @@
 public class IdResponse {
-    private int id;
+    private Long id;
 
-    public IdResponse(int id) {
+    public IdResponse(Long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
